regtest for domains.py
